Updated TTF file with the latest 4.0.3 version of the FontAwesome directly from the Website.
Also fixes an issue when adding an icon after updating the text of the button with setTitle:forState:UIControlStateNormal. It was duplicating the icon.

Updated the NSString category with the latest icon. A refactor was needed to update easyly the icon values and a better way to convert to FA string from an Enum value in HEX.

Deprecated unnecessary methods. Keeps the backwards compatibility of the Icon names.

The Enum has the same name convention than the CSS file on FA website, but keeps the older names for backward compatibility.